fastapi-backend/app/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# LangChain and LangGraph - Compatible versions
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.2.0

# Additional dependencies
httpx>=0.26.0
pydantic>=2.5.2